Fixing code ..
Multiple Github issues logged against the tutorial which features this notebook reveal an issue in the ordering of these code blocks. Changing the install block to occur before the declaration of variables addresses the bug. Without doing this, the second to last code block doesn't recognize the "storage_account" variable. Also, the uninstall step causes a blocking error that prevents the rest of the notebook from running. Recommend either troubleshooting that issue, or commenting that line out so that the sample can run.